requirements.txt
duckdb
requests
python-dotenv
thefuzz
python-Levenshtein
pandas